[SantokuApp] rename barcodeCharacter.ts to BarcodeCharacter.ts
