+++ example-app/SantokuApp/src/features/demo-barcode/types/BarcodeCharacter.ts
